refactor: `ExpireDateStatus`를 도메인 이동 후 문자열 리소스 매퍼 추가
